+++ internal/base/base.go
@@ -22,7 +22,7 @@ import (
 )
 
 // Version of asynq library and CLI.
-const Version = "0.19.1"
+const Version = "0.20.0"
 
 // DefaultQueueName is the queue name used if none are specified by user.
 const DefaultQueueName = "default"